feat: InputModalLayout 컴포넌트 추가
